// models/banner_data.ts
export interface IBannerData {
    mainBannerId: number;
    title: string;
    sort: number;
    pcImageUrl: string;
    mobileImageUrl: string;
    linkUrl: string;
    startDate: string;
    endDate: string;
    creator: string;
    updater: string;
    deleter: string | null;
    createdAt: string;
    updatedAt: string;
    deletedAt: string | null;
  }
  